# docs/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Factory Ops API
  description: |
    工廠值班工作管理系統 REST API(v1.3)。
    - API-first,Web 與未來 Native App 共用同一組 API。
    - 認證採 JWT Bearer Token(access + refresh)。JWT claims 帶 `userId`、`accountName`、`rootOrgId`、`orgPath[]`、`roles[]`、`groupIds[]`、`orgManagerScopes[]`(衍生自 Organization.managerId 反查)。
    - **多租戶隔離**:所有資源(除 `Organization` 自身)歸屬於某個 `rootOrgId`,server 強制 scope。跨 root org 操作僅 `ADMIN` 系統角色可進行。
    - **時間欄位**:wire format 使用 ISO 8601 + offset(例 `2026-05-04T08:30:00+08:00`,Q-17 拍板);儲存層為 UTC `Instant`。
    - 列表端點預設使用 cursor-based pagination,支援欄位投影與增量同步。
    - **Organization 為 polymorphic 樹**(FAB → DIVISION → DEPARTMENT → SECTION),只有 leaf type 可以承載 Group / Project / Task(見 ADR-0004)。每節點有 **單一 `managerId`**(對應 ORG_MANAGER)+ **多位 `leaderIds[]`**(業務負責人列表)— 見 ADR-0010。
    - **Group 為平面結構**(屬唯一 leaf Organization),polymorphic by `type`(DEFAULT / LINE / TEAM / SHIFT…);Group 加 `settings.qa` 用於設定 QA 雙簽機制(見 ADR-0011)。
    - **Template 機制**:支援 GLOBAL(ADMIN 維護)與 ORG(各 Org 維護)兩種 scope;可 fork from GLOBAL,可直接實例化(見 ADR-0006)。
    - **跨層派工(v1.3 改寫)**:任一 Organization 節點的 manager 可直接派 ActionRequest 到其子孫中的 leaf Org(`targetOrgId` 必為 leaf);**v1.3 移除多層 relay 流程**(見 ADR-0008 v1.3)。
    - **Task QA 雙簽**:Task 建立時從所屬 Project.groupIds 對應 Group settings.qa snapshot policy;之後 Group settings 變更不影響該 Task;Review action 走 `POST /tasks/{taskId}/review`(見 ADR-0011)。
    - **HR 整合**:User 為 HR User 的本地投影,以 `accountName` 對 HR Mock REST 查詢(見 ADR-0007 v1.3 Amendment)。
...
  version: 1.3.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /system/project-templates:
    get:
      tags: [SystemTemplates]
      summary: 列出 GLOBAL ProjectTemplate(所有角色可讀)
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: tag
          schema: { type: string }
          description: 以 tag 篩選(可重複)
        - in: query
          name: active
          schema: { type: boolean }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: GLOBAL ProjectTemplate 清單
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectTemplatePage' }
    post:
      tags: [SystemTemplates]
      summary: 建立 GLOBAL ProjectTemplate(ADMIN 限定)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProjectTemplateRequest' }
      responses:
        '201':
          description: 已建立
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectTemplate' }
        '403':
          $ref: '#/components/responses/Forbidden'
  /system/project-templates/{templateId}:
    parameters:
      - $ref: '#/components/parameters/TemplateIdParam'
    get:
      tags: [SystemTemplates]
      summary: 讀取 GLOBAL ProjectTemplate
      parameters:
        - in: query
          name: version
          schema: { type: integer }
      responses:
        '200':
          description: ProjectTemplate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectTemplate' }
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [SystemTemplates]
      summary: 更新 GLOBAL ProjectTemplate(ADMIN;產生新 version)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProjectTemplateRequest' }
      responses:
        '200':
          description: 新 version 已建立
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectTemplate' }
    delete:
      tags: [SystemTemplates]
      summary: Deactivate GLOBAL ProjectTemplate(ADMIN)
      responses:
        '204': { description: No Content }
  /system/task-templates:
    get:
      tags: [SystemTemplates]
      summary: 列出 GLOBAL TaskTemplate
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: tag
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: active
          schema: { type: boolean }
      responses:
        '200':
          description: GLOBAL TaskTemplate 清單
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskTemplatePage' }
    post:
      tags: [SystemTemplates]
      summary: 建立 GLOBAL TaskTemplate(ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTaskTemplateRequest' }
      responses:
        '201':
          description: 已建立
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskTemplate' }
  /system/task-templates/{templateId}:
    parameters:
      - $ref: '#/components/parameters/TemplateIdParam'
    get:
      tags: [SystemTemplates]
      summary: 讀取 GLOBAL TaskTemplate
      parameters:
        - in: query
          name: version
          schema: { type: integer }
      responses:
        '200':
          description: TaskTemplate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskTemplate' }
    patch:
      tags: [SystemTemplates]
      summary: 更新 GLOBAL TaskTemplate(ADMIN;產生新 version)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTaskTemplateRequest' }
      responses:
        '200':
          description: 新 version 已建立
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskTemplate' }
    delete:
      tags: [SystemTemplates]
      summary: Deactivate GLOBAL TaskTemplate(ADMIN)
      responses:
        '204': { description: No Content }

  # ===================== Fork from Global to ORG =====================
  /orgs/{orgId}/project-templates/fork-from-global/{globalTplId}:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
      - in: path
        name: globalTplId
        required: true
        schema: { type: string }
    post:
      tags: [ProjectTemplates]
      summary: 從 GLOBAL ProjectTemplate Fork 為 ORG-scoped(獨立 version chain)
      description: |
        將指定 GLOBAL ProjectTemplate 深拷貝為當前 root org 的 ORG ProjectTemplate(version=1)。
        Fork 後完全脫鉤,GLOBAL 後續演進不影響此 ORG 副本。
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string, description: ORG 副本的 code(預設沿用 GLOBAL code) }
                name: { type: string, description: 覆寫名稱 }
      responses:
        '201':
          description: 已 fork
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectTemplate' }
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /orgs/{orgId}/task-templates/fork-from-global/{globalTplId}:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
      - in: path
        name: globalTplId
        required: true
        schema: { type: string }
    post:
      tags: [TaskTemplates]
      summary: 從 GLOBAL TaskTemplate Fork 為 ORG-scoped
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
                name: { type: string }
      responses:
        '201':
          description: 已 fork
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskTemplate' }

  # ===================== ProjectTemplates (ORG) =====================
  /orgs/{orgId}/project-templates:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    get:
      tags: [ProjectTemplates]
      summary: 列出 ORG ProjectTemplate(同 root org)
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: tag
          schema: { type: string }
        - in: query
          name: active
          schema: { type: boolean }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: includeGlobal
          schema: { type: boolean, default: false }
          description: 是否同時回傳 GLOBAL Templates(GROUP_ADMIN 找範本時方便)
      responses:
        '200':
          description: ProjectTemplate 清單
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectTemplatePage' }
    post:
      tags: [ProjectTemplates]
      summary: 建立 ProjectTemplate(version=1)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProjectTemplateRequest' }
      responses:
        '201':
          description: 已建立
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectTemplate' }
        '422':
          $ref: '#/components/responses/ValidationError'
  /orgs/{orgId}/project-templates/{templateId}:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
      - $ref: '#/components/parameters/TemplateIdParam'
    get:
      tags: [ProjectTemplates]
      summary: 讀取 ProjectTemplate
      parameters:
        - in: query
          name: version
          schema: { type: integer }
          description: 指定版本(預設最新)
      responses:
        '200':
          description: ProjectTemplate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectTemplate' }
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [ProjectTemplates]
      summary: 更新 ProjectTemplate(產生新 version)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProjectTemplateRequest' }
      responses:
        '200':
          description: 新 version 已建立
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectTemplate' }
    delete:
      tags: [ProjectTemplates]
      summary: Deactivate ProjectTemplate(軟性,設 active=false)
      responses:
        '204': { description: No Content }

  # ===================== TaskTemplates =====================
  /orgs/{orgId}/task-templates:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    get:
      tags: [TaskTemplates]
      summary: 列出 ORG TaskTemplate(同 root org)
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: tag
          schema: { type: string }
        - in: query
          name: active
          schema: { type: boolean }
        - in: query
          name: includeGlobal
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: TaskTemplate 清單
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskTemplatePage' }
    post:
      tags: [TaskTemplates]
      summary: 建立 TaskTemplate(version=1)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTaskTemplateRequest' }
      responses:
        '201':
          description: 已建立
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskTemplate' }
        '422':
          $ref: '#/components/responses/ValidationError'
  /orgs/{orgId}/task-templates/{templateId}:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
      - $ref: '#/components/parameters/TemplateIdParam'
    get:
      tags: [TaskTemplates]
      summary: 讀取 TaskTemplate
      parameters:
        - in: query
          name: version
          schema: { type: integer }
      responses:
        '200':
          description: TaskTemplate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskTemplate' }
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [TaskTemplates]
      summary: 更新 TaskTemplate(產生新 version)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTaskTemplateRequest' }
      responses:
        '200':
          description: 新 version 已建立
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskTemplate' }
    delete:
      tags: [TaskTemplates]
      summary: Deactivate TaskTemplate(設 active=false)
      responses:
        '204': { description: No Content }

  # ===================== Users =====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrgIdParam:
      in: path
      name: orgId
      required: true
      schema: { type: string }
      description: |
        Organization ID。在 `/orgs/{orgId}/...` 子路徑中,根據端點語意可能是 root org 或 任一節點:
        - `/orgs/{orgId}/groups`、`/orgs/{orgId}/project-templates` 等管理面 API:`orgId` 為 **rootOrgId**
        - `/orgs/{orgId}/dispatch-action-request`、`/orgs/{orgId}/transfer-leader`:`orgId` 為**特定節點**(可能任一層)
...
    TemplateIdParam:
      in: path
      name: templateId
      required: true
      schema: { type: string }
...
    CursorParam:
      in: query
      name: cursor
      schema: { type: string }
      description: cursor-based pagination 游標
    LimitParam:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
...
    Forbidden:
      description: 無權限
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: 找不到資源
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: 業務規則衝突(狀態流、唯一性)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: 輸入驗證失敗
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
  schemas:
...
    Problem:
      description: RFC 7807 problem details
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
      required: [title, status]
...
    PageInfo:
      type: object
      properties:
        nextCursor: { type: string, nullable: true }
        hasMore: { type: boolean }
...
    TemplateScope:
      type: string
      enum: [GLOBAL, ORG]
      description: GLOBAL 由 ADMIN 維護,所有 Org 可見;ORG 屬某 root org
...
    Priority:
      type: string
      enum: [LOW, NORMAL, HIGH, URGENT]
...
    ProjectTemplate:
      type: object
      properties:
        id: { type: string }
        scope: { $ref: '#/components/schemas/TemplateScope' }
        rootOrgId:
          type: string
          nullable: true
          description: GLOBAL scope 時為 null;ORG scope 時必填
        name: { type: string }
        code: { type: string, description: 同 (scope, rootOrgId) 內 unique }
        descriptionMarkdown: { type: string }
        defaultMemberRoles:
          type: object
          description: "角色配置(例 { ENGINEER: 1, OPERATOR: 2 })"
          additionalProperties:
            type: integer
            minimum: 0
        taskTemplateRefs:
          type: array
          items: { $ref: '#/components/schemas/TaskTemplateRef' }
        estimatedDurationDays: { type: integer, minimum: 0 }
        tags:
          type: array
          items: { type: string }
        forkedFrom:
          allOf:
            - $ref: '#/components/schemas/ForkRef'
          nullable: true
          description: 若由 GLOBAL fork 而來則記錄
        version: { type: integer, minimum: 1 }
        active: { type: boolean }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
        schemaVersion: { type: integer }
      required: [id, scope, name, code, version, active]
    TaskTemplateRef:
      type: object
      required: [taskTemplateId, version, sortOrder]
      properties:
        taskTemplateScope: { $ref: '#/components/schemas/TemplateScope' }
        taskTemplateId: { type: string }
        version: { type: integer, minimum: 1 }
        sortOrder: { type: integer }
        estimatedHours: { type: number, format: float }
    ForkRef:
      type: object
      required: [sourceTemplateId, sourceVersion, forkedAt]
      properties:
        sourceTemplateId: { type: string }
        sourceVersion: { type: integer }
        forkedAt: { type: string, format: date-time }
        forkedBy: { type: string }
    CreateProjectTemplateRequest:
      type: object
      required: [name, code]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        code: { type: string, minLength: 1, maxLength: 64 }
        descriptionMarkdown: { type: string }
        defaultMemberRoles:
          type: object
          additionalProperties: { type: integer, minimum: 0 }
        taskTemplateRefs:
          type: array
          items: { $ref: '#/components/schemas/TaskTemplateRef' }
        estimatedDurationDays: { type: integer, minimum: 0 }
        tags:
          type: array
          items: { type: string }
    UpdateProjectTemplateRequest:
      type: object
      description: 更新會建立新 version,既有 version 凍結
      properties:
        name: { type: string }
        descriptionMarkdown: { type: string }
        defaultMemberRoles:
          type: object
          additionalProperties: { type: integer, minimum: 0 }
        taskTemplateRefs:
          type: array
          items: { $ref: '#/components/schemas/TaskTemplateRef' }
        estimatedDurationDays: { type: integer }
        tags:
          type: array
          items: { type: string }
        active: { type: boolean }
    ProjectTemplatePage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ProjectTemplate' }
        pageInfo: { $ref: '#/components/schemas/PageInfo' }
    TaskTemplate:
      type: object
      properties:
        id: { type: string }
        scope: { $ref: '#/components/schemas/TemplateScope' }
        rootOrgId:
          type: string
          nullable: true
          description: GLOBAL scope 時為 null;ORG scope 時必填
        name: { type: string }
        code: { type: string }
        type:
          type: string
          description: 對應 Task `type`(必須是已註冊的 type)
        defaultPriority: { $ref: '#/components/schemas/Priority' }
        defaultAttributes:
          type: object
          additionalProperties: true
          description: 預填 Task attributes,實例化時 clone
        descriptionMarkdownTemplate: { type: string }
        defaultChecklist:
          type: array
          items: { $ref: '#/components/schemas/ChecklistItem' }
        attachmentHints:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        forkedFrom:
          allOf:
            - $ref: '#/components/schemas/ForkRef'
          nullable: true
        version: { type: integer, minimum: 1 }
        active: { type: boolean }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
        schemaVersion: { type: integer }
      required: [id, scope, name, code, type, version, active]
...
    CreateTaskTemplateRequest:
      type: object
      required: [name, code, type]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        code: { type: string, minLength: 1, maxLength: 64 }
        type: { type: string }
        defaultPriority: { $ref: '#/components/schemas/Priority' }
        defaultAttributes:
          type: object
          additionalProperties: true
        descriptionMarkdownTemplate: { type: string }
        defaultChecklist:
          type: array
          items: { $ref: '#/components/schemas/ChecklistItem' }
        attachmentHints:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
    UpdateTaskTemplateRequest:
      type: object
      description: 更新會建立新 version,既有 version 凍結
      properties:
        name: { type: string }
        defaultPriority: { $ref: '#/components/schemas/Priority' }
        defaultAttributes:
          type: object
          additionalProperties: true
        descriptionMarkdownTemplate: { type: string }
        defaultChecklist:
          type: array
          items: { $ref: '#/components/schemas/ChecklistItem' }
        attachmentHints:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        active: { type: boolean }
    TaskTemplatePage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/TaskTemplate' }
        pageInfo: { $ref: '#/components/schemas/PageInfo' }

    # ---------- Comment ----------
...